backend/requirements.txt
# Core dependencies
fastapi
uvicorn
motor
pymongo
pydantic
passlib[bcrypt]
PyJWT
python-multipart
requests
whisper
transformers
torchaudio

# AI Pothole Detection dependencies
opencv-python-headless
Pillow
numpy

# Optional: ML dependencies (large downloads, not required for basic functionality)
# Install these for full NLP/Whisper capabilities:
# openai-whisper
# torch
# transformers
# spacy
# After installing spacy, run: python -m spacy download en_core_web_sm
